setup image compression

+++ ml_proxy/src/rpc.h
@@ -7,6 +7,6 @@
 #ifndef ML_RPC
 #define ML_RPC
 
-int ml_rpc_getImage(RequestStatus* status, char** img_buffer, size_t* img_length);
+int ml_rpc_getImage(RequestStatus* status, char** img_buffer, int* img_length);
 
 #endif
